requirement.txt
streamlit
requests            
pandas      
python-dotenv       
langchain
langchain_core
langchain-groq    
langgraph  
fastapi[all]
uvicorn
SQLAlchemy
onnxruntime-gpu
onnx
opencv-python 
scipy
rfdetr
filterpy
onnxscript
plotly
PyYAML